Move handleMessage.js to become a middleware
